Extract `AngularDragDropService` to a separate class for better modularity
