+++ zeppelin-web-angular/src/app/services/public-api.ts
@@ -10,6 +10,7 @@
  * limitations under the License.
  */
 
+export * from './angular-drag-drop.service';
 export * from './array-ordering.service';
 export * from './base-url.service';
 export * from './completion.service';
@@ -20,11 +21,11 @@ export * from './interpreter.service';
 export * from './job-manager.service';
 export * from './message.service';
 export * from './ng-z.service';
+export * from './note-action.service';
 export * from './note-list.service';
 export * from './note-status.service';
 export * from './note-var-share.service';
 export * from './notebook-repos.service';
-export * from './notebook.service';
 export * from './runtime-compiler.service';
 export * from './save-as.service';
 export * from './security.service';
